05-18 목요일

+++ frontend/src/components/custom/errorHandlingMiddleware.js
@@ -1,8 +1,13 @@
 export const errorHandlingMiddleware = (error) => {
+  console.log(error);
+  if (error === 'TypeError: Failed to fetch') {
+    window.location.href = '/';
+    return;
+  }
   // 에러 처리 로직 작성
   alert(error);
   // 추가적인 에러 처리 로직을 이곳에 작성 가능
   // 예를 들어, 사용자에게 에러 메시지를 표시하거나 로깅 등을 수행할 수 있음
   localStorage.clear();
-  window.location.href = "/";
+  window.location.href = '/';
 };
